refactor(frontend): Add shared helper module for assignments
