+++ frontend/src/app/assignment/modules/shared/author-name/author-name.component.ts
