+++ frontend/src/app/assignment/modules/shared/pipes/task-color.pipe.spec.ts
