+++ frontend/src/app/assignment/modules/shared/task-list/task-list.component.ts
@@ -1,6 +1,6 @@
 import {Component, Input} from '@angular/core';
-import {CreateEvaluationDto} from '../../model/evaluation';
-import Task from '../../model/task';
+import {CreateEvaluationDto} from '../../../model/evaluation';
+import Task from '../../../model/task';
 
 @Component({
   selector: 'app-task-list',
